pkg: move loop interfaces to pkg/types to unify type set
