+++ pkg/loop/relayer_service.go
@@ -22,7 +22,7 @@ type RelayerService struct {
 
 // NewRelayerService returns a new [*RelayerService].
 // cmd must return a new exec.Cmd each time it is called.
-func NewRelayerService(lggr logger.Logger, grpcOpts GRPCOpts, cmd func() *exec.Cmd, config string, keystore Keystore) *RelayerService {
+func NewRelayerService(lggr logger.Logger, grpcOpts GRPCOpts, cmd func() *exec.Cmd, config string, keystore types.Keystore) *RelayerService {
 	newService := func(ctx context.Context, instance any) (Relayer, error) {
 		plug, ok := instance.(PluginRelayer)
 		if !ok {
